refactor: load sections independently
